+++ sql/driving_distance/withPointsDD.sql
@@ -39,7 +39,7 @@ CREATE FUNCTION pgr_withPointsDD(
     directed BOOLEAN DEFAULT true,
     details BOOLEAN DEFAULT false,
 
-    OUT seq INTEGER,
+    OUT seq BIGINT,
     OUT depth BIGINT,
     OUT start_vid BIGINT,
     OUT node BIGINT,
@@ -68,7 +68,7 @@ CREATE FUNCTION pgr_withPointsDD(
     details BOOLEAN DEFAULT false,
     equicost BOOLEAN DEFAULT false,
 
-    OUT seq INTEGER,
+    OUT seq BIGINT,
     OUT depth BIGINT,
     OUT start_vid BIGINT,
     OUT node BIGINT,
